Refactor collateral fetching into separate module
